REFACTOR tests
